requirements.txt
# Google cloud platform
google-python-cloud-debugger
google-cloud-logging
google-cloud-tasks==2.7.1
google-cloud-firestore

# Google cloud functions
functions-framework==3.*
Flask==2.0.2

# TG bot + functions
gunicorn
requests
pydantic
cachetools
aiohttp[speedups]
aiocache
git+https://github.com/jazzband/prettytable#egg=prettytable

# cf_update
aiocron
aiotinydb
tinydb==3.12.1
ujson
python-telegram-bot==20.0a2
